asdasd

+++ include/ecs.h
@@ -1,12 +1,11 @@
 #pragma once
-#include <stdbool.h>
 #include <string.h>
-#include <stdint.h>
+#include <stdlib.h>
 
 typedef struct {
-  uint16_t type;
+  unsigned int type;
   void *content;
 } component_t;
 
-component_t _component(uint16_t type, void *data, unsigned long size);
+component_t _component(unsigned int type, void *data, unsigned long size);
 #define COMPONENT(datatype, type, data) _component(type, &data, sizeof(datatype))
