+++ src/ecs.c
@@ -1,8 +1,6 @@
 #include "ecs.h"
-#include <stdlib.h>
-#include <stdio.h>
 
-component_t _component(uint16_t type, void *data, unsigned long size) {
+component_t _component(unsigned int type, void *data, unsigned long size) {
   component_t _c = (component_t) {
     type,
     NULL,
